Refactor : migrate web3 config
