+++ back/src/main/java/com/undefined/undefined/global/web3/klaytn/event/events/ChooseWinnerEvent.java
@@ -1,8 +1,7 @@
-package com.undefined.undefined.global.web3.klaytn.event;
+package com.undefined.undefined.global.web3.klaytn.event.events;
 
 import com.undefined.undefined.global.web3.klaytn.dto.ChooseWinnerDto;
 import com.undefined.undefined.global.web3.klaytn.mapper.EventTypeMapper;
-import org.springframework.stereotype.Component;
 import org.web3j.abi.EventEncoder;
 import org.web3j.abi.FunctionReturnDecoder;
 import org.web3j.abi.TypeReference;
