+++ back/src/main/java/com/undefined/undefined/global/web3/klaytn/event/events/ClaimAllNftsEvent.java
@@ -1,6 +1,5 @@
-package com.undefined.undefined.global.web3.klaytn.event;
+package com.undefined.undefined.global.web3.klaytn.event.events;
 
-import org.springframework.stereotype.Component;
 import org.web3j.abi.EventEncoder;
 import org.web3j.abi.FunctionReturnDecoder;
 import org.web3j.abi.TypeReference;
