+++ back/src/main/java/com/undefined/undefined/global/web3/klaytn/event/handler/BuyTicketsHandler.java
@@ -1,8 +1,8 @@
-package com.undefined.undefined.global.web3.klaytn.handler;
+package com.undefined.undefined.global.web3.klaytn.event.handler;
 
 import com.undefined.undefined.domain.ticket.service.TicketService;
 import com.undefined.undefined.global.web3.klaytn.dto.BuyTicketsDto;
-import com.undefined.undefined.global.web3.klaytn.event.BuyTicketsEvent;
+import com.undefined.undefined.global.web3.klaytn.event.events.BuyTicketsEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
