+++ back/src/main/java/com/undefined/undefined/global/web3/klaytn/event/handler/ClaimBalanceHandler.java
@@ -1,8 +1,8 @@
-package com.undefined.undefined.global.web3.klaytn.handler;
+package com.undefined.undefined.global.web3.klaytn.event.handler;
 
 import com.undefined.undefined.domain.raffle.service.RaffleService;
 import com.undefined.undefined.global.web3.klaytn.dto.ClaimBalanceDto;
-import com.undefined.undefined.global.web3.klaytn.event.ClaimBalanceEvent;
+import com.undefined.undefined.global.web3.klaytn.event.events.ClaimBalanceEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
