+++ back/src/main/java/com/undefined/undefined/global/web3/klaytn/event/handler/ClaimNftHandler.java
@@ -1,8 +1,8 @@
-package com.undefined.undefined.global.web3.klaytn.handler;
+package com.undefined.undefined.global.web3.klaytn.event.handler;
 
 import com.undefined.undefined.domain.raffle.service.RaffleService;
 import com.undefined.undefined.global.web3.klaytn.dto.ClaimNftDto;
-import com.undefined.undefined.global.web3.klaytn.event.ClaimNftEvent;
+import com.undefined.undefined.global.web3.klaytn.event.events.ClaimNftEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
